pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "cohere>=5.14.0",
    "fastapi>=0.115.10",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "motor>=3.7.0",
    "openai>=1.65.5",
    "psycopg2-binary>=2.9.10",
    "pydantic-mongo>=2.3.0",
    "pydantic-settings>=2.8.1",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.13.3",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
